Move common components to components folder; create hover effect for images; fix height styling; add button to image
