+++ src/panels/SkyBrowserPanel/components/IconImage.module.css
@@ -1,0 +1,3 @@
+.iconImage:hover {
+  filter: brightness(1.3);
+}
