+++ src/panels/SkyBrowserPanel/components/ImageInfoPopover.tsx
